Add: PersonaController y cambios de int a Long
Se agregó el controller de persona que incluye el registro de empleado y de cliente

+++ src/main/java/com/oo2/grupo4/entities/Empleado.java
@@ -22,7 +22,7 @@ import lombok.ToString;
 public class Empleado extends Persona{
 
 	private String rol;
-	private long legajo;
+	private int legajo;
 	
 	@ManyToOne(cascade = CascadeType.ALL)
 	@JoinColumn(name = "area_id_area")
