+++ src/main/java/com/oo2/grupo4/services/interfaces/IClienteService.java
@@ -4,7 +4,7 @@ import com.oo2.grupo4.entities.Cliente;
 
 public interface IClienteService {
 
-    Cliente crearCliente(String nombre, String apellido, int dni, String nroCliente);
+    Cliente crearCliente(String nombre, String apellido, Long dni, String nroCliente);
     
     void completarCliente(int idPersona, String nroCliente);
